+++ src/main/java/br/com/zalf/prolog/webservice/entrega/relatorio/RelatorioEntregaService.java
@@ -122,4 +122,36 @@ public class RelatorioEntregaService {
             return null;
         }
     }
+
+    public void getExtratoMapasIndicadorCsv(Long codEmpresa, String codRegional, String codUnidade, String cpf,
+                                            Long dataInicial, Long dataFinal, String equipe, OutputStream out) {
+        try {
+            dao.getExtratoMapasIndicadorCsv(codEmpresa, codRegional, codUnidade, cpf, new Date(dataInicial), new Date(dataFinal), equipe, out);
+        } catch (SQLException | IOException e) {
+            Log.e(TAG, String.format("Erro ao gerar o relatório que esrtatifica os indicadores de cada mapa. \n" +
+                    "codEmpresa: %d \n" +
+                    "codRegional: %s \n" +
+                    "codUnidade: %s \n" +
+                    "equipe: %s \n" +
+                    "cpf: %d \n" +
+                    "Período: %d a %d", codEmpresa, codRegional, codUnidade, equipe, cpf, dataInicial, dataFinal), e);
+            throw new RuntimeException();
+        }
+    }
+
+    public Report getExtratoMapasIndicadorReport(Long codEmpresa, String codRegional, String codUnidade, String cpf,
+                                                 Long dataInicial, Long dataFinal, String equipe) {
+        try {
+            return dao.getExtratoMapasIndicadorReport(codEmpresa, codRegional, codUnidade, cpf, new Date(dataInicial), new Date(dataFinal), equipe);
+        } catch (SQLException e) {
+            Log.e(TAG, String.format("Erro ao gerar o relatório que esrtatifica os indicadores de cada mapa. \n" +
+                    "codEmpresa: %d \n" +
+                    "codRegional: %s \n" +
+                    "codUnidade: %s \n" +
+                    "equipe: %s \n" +
+                    "cpf: %d \n" +
+                    "Período: %d a %d", codEmpresa, codRegional, codUnidade, equipe, cpf, dataInicial, dataFinal), e);
+            throw new RuntimeException();
+        }
+    }
 }